+++ src/Form/LegacyConsumer/templates/label.html.php
@@ -1,5 +1,5 @@
 <?php /** @var Give\Framework\FieldsAPI\Field $field */ ?>
 <?php /** @var string $fieldIdAttribute */ ?>
 <label class="give-label" for="<?php echo $fieldIdAttribute; ?>">
-	<?php include plugin_dir_path( __FILE__ ) . '/label-content.html.php'; ?>
+	<?php include plugin_dir_path( __FILE__ ) . 'label-content.html.php'; ?>
 </label>