+++ src/Form/LegacyConsumer/templates/radio.html.php
@@ -1,19 +1,13 @@
 <?php /** @var Give\Framework\FieldsAPI\Radio $field */ ?>
 <?php /** @var string $fieldIdAttribute */ ?>
+<?php /** @var string $visibilityConditionsAttribute */ ?>
 <?php /* Fieldsets + legends are terrible to style, so we just use the semantic markup and style something else. */ ?>
-<fieldset
-	<?php
-	if ( $conditions = $field->getVisibilityConditions() ) {
-		$conditions = esc_attr( json_encode( $conditions ) );
-		echo "data-field-visibility-conditions=\"$conditions\"";
-	}
-	?>
->
+<fieldset <?php echo $visibilityConditionsAttribute ?>>
 	<legend class="screen-reader-text">
-		<?php include plugin_dir_path( __FILE__ ) . '/label-content.html.php'; ?>
+		<?php include plugin_dir_path( __FILE__ ) . 'label-content.html.php'; ?>
 	</legend>
 	<div class="give-label" aria-hidden="true">
-		<?php include plugin_dir_path( __FILE__ ) . '/label-content.html.php'; ?>
+		<?php include plugin_dir_path( __FILE__ ) . 'label-content.html.php'; ?>
 	</div>
 	<?php foreach ( $field->getOptions() as $index => $option ) : ?>
 		<?php $id = $fieldIdAttribute . '-' . $index; ?>